feat: Phi4分析機能のリファクタリングと新機能の追加 - `Phi4Analysis` コンポーネントをリファクタリングし、分析結果の表示を `AnalysisForm` コンポーネントに分割 - 新しいフック `usePhi4Analysis` を追加し、状態管理とAPI呼び出しを整理 - `package.json` と `package-lock.json` の依存関係を更新 - 不要な `Phi4Analysis.tsx` ファイルを削除 - E2Eテストを強化し、不正なファイルタイプのエラーハンドリングを追加
