+++ frontend/package.json
@@ -23,7 +23,9 @@
     "react-dom": "^18.3.1",
     "react-router-dom": "^7.1.3",
     "react-scripts": "^5.0.1",
-    "sass-embedded": "^1.83.1"
+    "sass": "^1.86.3",
+    "sync-child-process": "^1.0.2",
+    "sync-message-port": "^1.1.3"
   },
   "devDependencies": {
     "@eslint/js": "^9.17.0",
@@ -35,7 +37,7 @@
     "eslint-plugin-react-hooks": "^5.0.0",
     "eslint-plugin-react-refresh": "^0.4.16",
     "globals": "^15.14.0",
-    "sass-embedded": "^1.83.1",
+    "sass-embedded": "^1.86.3",
     "typescript": "^4.9.5",
     "typescript-eslint": "^8.18.2",
     "vite": "^6.0.5"
